Add YDL_RAW_OPTIONS config variable to pass ydl vars
YDL_RAW_OPTIONS is a dict that can be set to pass raw parameters to
youtube-dl.
In the default config, ignoreerrors has been added and set to True as a
default element for YDL_RAW_OPTIONS. This allows playslist downloads to
go on if an error is encountered on a video.